Avoid cyclic import issues, when we need to initialize `WorkspaceContext` from `AccountContext` (#1611)
This PR splits `cli_command.py` into `workspace_cli.py` and
`account_cli.py`
