+++ .github/workflows/no-cheat.yml
@@ -3,10 +3,15 @@ name: no-cheat
 on:
   pull_request:
     types: [opened, synchronize]
+  push:
+    # required for merge queue to work. jobs.integration.if will mark it as skipped
+    branches:
+      - main
 
 jobs:
   no-pylint-disable:
     runs-on: ubuntu-latest
+    if: github.event_name == 'pull_request' && (github.event.action == 'opened' || github.event.action == 'synchronize')
     steps:
       - name: Checkout
         uses: actions/checkout@v4
