+++ tests/unit/test_cli.py
@@ -46,7 +46,8 @@ from databricks.labs.ucx.cli import (
     migrate_tables,
     create_missing_principals,
 )
-from databricks.labs.ucx.contexts.cli_command import WorkspaceContext, AccountContext
+from databricks.labs.ucx.contexts.account_cli import AccountContext
+from databricks.labs.ucx.contexts.workspace_cli import WorkspaceContext
 from databricks.labs.ucx.hive_metastore import TablesCrawler
 from databricks.labs.ucx.hive_metastore.tables import Table
